fix(kwb_modbus): restore v1.2.0 changes lost by first HACS sync, drop pymodbus pin
The first successful HACS sync run overwrote the HACS repo's v1.2.0
(discovery fix for indexed modules, German expert_mode translations,
pymodbus==3.11.2 pin) with this repo's stale copy, since the sync is a
one-way rsync --delete from here to there. Restoring v1.2.0's content
here and keeping requirements empty (see previous commit) so the next
sync carries the fix forward instead of regressing the HACS repo again.

+++ custom_components/kwb_modbus/number.py
@@ -1,7 +1,5 @@
 """Number platform for writable KWB Modbus holding registers."""
 
-from __future__ import annotations
-
 from homeassistant.components.number import NumberEntity, NumberMode
 from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
@@ -145,7 +143,7 @@ class KWBNumberEntity(CoordinatorEntity[KWBDataUpdateCoordinator], NumberEntity)
     async def async_set_native_value(self, value: float) -> None:
         """Write numeric value to holding register(s)."""
         scale = self._register.scale or 1.0
-        raw = int(round(value / scale))
+        raw = round(value / scale)
 
         success = False
         if self._register.count == 1:
